+++ server/cache_test.go
@@ -279,13 +279,13 @@ func (s *testClusterInfoSuite) testStoreHeartbeat(c *C, cache *clusterInfo) {
 		c.Assert(cache.getStoreCount(), Equals, int(i+1))
 
 		stats := store.stats
-		c.Assert(stats.LeaderCount, Equals, 0)
+		c.Assert(stats.LeaderCount, Equals, uint32(0))
 		c.Assert(stats.LastHeartbeatTS.IsZero(), IsTrue)
 
 		c.Assert(cache.handleStoreHeartbeat(storeStats), IsNil)
 
 		stats = cache.getStore(store.GetId()).stats
-		c.Assert(stats.LeaderCount, Equals, 1)
+		c.Assert(stats.LeaderCount, Equals, uint32(1))
 		c.Assert(stats.LastHeartbeatTS.IsZero(), IsFalse)
 	}
